Update Task1.php
Task 1:
$quarter = [ ['start' => '12-01', 'ends' => '02-28'],
        ['start' => '03-01', 'ends' => '05-31'], 
        ['start' => '06-01', 'ends' => '08-31'], 
        ['start' => '09-01', 'ends' => '11-31']]; 
Every month belongs to one quarter. Kindly develop the function to find out which given month belongs to which quarter.

Solution Description:
I gave two solutions for this problem one for static input of quarters and the other is for dynamic input of quarters which are specified in get_Quarter_Static.php and get_Quarter_Dynamic.php respectively. Task1.php file accepts the input and sends it another file to get the solution.